WO#6 fix: cast bf16 -> fp32 before numpy in dormant '=' capture
TransformerLens LayerNormPre.forward re-casts its output back to cfg.dtype
(bfloat16), so the .float() before ln_final was undone; numpy has no bf16
dtype -> TypeError on .numpy() at the '=' residual capture. Add .float()
after ln_final. Path-patch half (head_path_patch_*.json) was unaffected and
resumes from its checkpoint. CPU gates green (test_wo_logic, test_wo6_mock).

+++ cells/82s_wo_pathpatch_dormant.py
@@ -220,8 +220,8 @@ def _dorm_capture_eq(tag, bundle):
     def _eq_resid(cache, eq):
         rstack = torch.stack([cache[f"blocks.{L}.hook_resid_post"][0, eq] for L in range(nL)]).float()
         if _has_ln:
-            rstack = model.ln_final(rstack)
-        return rstack.detach().cpu().numpy().astype(np.float16)
+            rstack = model.ln_final(rstack)   # TL LayerNormPre re-casts output to cfg.dtype (bf16)
+        return rstack.detach().float().cpu().numpy().astype(np.float16)  # fp32 first: numpy has no bf16
 
     seen, rows, prod, ft = set(), [], [], []
     for (B, C) in dorm_pairs:
